requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altair==4.2.2
annotated-types==0.6.0
anyio==3.6.2
arabic-reshaper==3.0.0
argilla==1.7.0
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asgiref==3.7.2
asn1crypto==1.5.1
astroid==3.0.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==23.1.0
awsebcli==3.20.10
backcall==0.2.0
backoff==2.2.1
beautifulsoup4==4.12.3
black==23.3.0
bleach==6.0.0
blinker==1.6.2
boto3==1.28.23
botocore==1.31.23
Brotli==1.1.0
bs4==0.0.1
cachetools==5.3.0
cement==2.8.2
certifi==2023.11.17
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.1.0
chromadb==0.3.26
click==8.1.3
clickhouse-connect==0.6.4
colorama==0.4.3
coloredlogs==15.0.1
comm==0.1.3
commonmark==0.9.1
contourpy==1.0.7
cryptography==41.0.7
cssselect2==0.7.0
cycler==0.11.0
dataclasses-json==0.5.9
debugpy==1.6.7
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dill==0.3.6
distlib==0.3.8
dj-database-url==2.1.0
Django==4.2.3
django-admin==2.0.2
django-adminlte-3==0.1.6
django-appconf==1.0.5
django-bitfield==2.2.0
django-browser-reload==1.10.0
django-compressor==4.3.1
django-compressor-postcss==0.8.2
django-cors-headers==4.3.1
django-excel-response2==3.0.6
django-filter==23.2
django-geo==0.8
django-heroku==0.3.1
django-jet==1.0.8
django-js-asset==2.2.0
django-mptt==0.16.0
django-ninja==1.1.0
django-separatedvaluesfield==0.6.0
django-six==1.0.5
django-tailwind==3.6.0
django-timezone-field==6.1.0
django-uuidfield==0.5.0
django-widget-tweaks==1.5.0
djangorestframework==3.14.0
djangorestframework-csv==2.1.1
djangorestframework-simplejwt==5.3.1
duckdb==0.8.1
entrypoints==0.4
et-xmlfile==1.1.0
excel-base==1.0.4
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.95.1
fastjsonschema==2.16.3
filelock==3.13.1
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.39.3
fqdn==1.5.1
frozenlist==1.3.3
gitdb==4.0.10
GitPython==3.1.31
greenlet==2.0.2
gunicorn==21.2.0
h11==0.14.0
hnswlib==0.7.0
html5lib==1.1
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
humanfriendly==10.0
idna==3.6
importlib-metadata==6.6.0
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
isort==5.12.0
isoweek==1.3.3
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.2.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
langchain==0.0.225
langchainplus-sdk==0.0.20
lazy-object-proxy==1.9.0
libmagic==1.0
lxml==5.1.0
lz4==4.3.2
Markdown==3.4.3
MarkupSafe==2.1.3
marshmallow==3.19.0
marshmallow-enum==1.5.1
matplotlib==3.7.1
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==2.0.5
monotonic==1.6
mpmath==1.3.0
mrjob==0.7.4
msg-parser==1.2.0
multidict==6.0.4
mypy-extensions==1.0.0
mysql-connector-python==8.0.33
mysqlclient==2.2.0
nbclassic==0.5.6
nbclient==0.7.4
nbconvert==7.3.1
nbformat==5.8.0
nest-asyncio==1.5.6
nltk==3.8.1
notebook==6.5.4
notebook_shim==0.2.3
numexpr==2.8.4
numpy==1.24.4
oauthlib==3.2.2
olefile==0.46
onnxruntime==1.15.1
openai==0.27.4
openapi-schema-pydantic==1.2.4
openpyxl==3.1.2
oscrypto==1.3.0
overrides==7.3.1
packaging==23.2
pandas==2.0.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.10.1
pdfminer.six==20221105
pickleshare==0.7.5
pillow==10.2.0
platformdirs==4.2.0
postgis==1.0.4
posthog==3.0.1
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
psycopg2==2.9.7
psycopg2-binary==2.9.9
pulsar-client==3.2.0
pure-eval==0.2.2
pyarrow==12.0.0
pycparser==2.21
pydantic==2.6.1
pydantic_core==2.16.2
pydeck==0.8.1b0
pydyf==0.8.0
Pygments==2.15.1
pyHanko==0.21.0
pyhanko-certvalidator==0.26.3
PyJWT==2.8.0
pylint==3.0.3
Pympler==1.0.1
PyMySQL==1.0.3
pypandoc==1.11
pyparsing==3.0.9
pypdf==4.0.2
PyPDF2==3.0.1
pyphen==0.14.0
pypiwin32==223
pypng==0.20220715.0
pyreadline3==3.4.1
pyrsistent==0.19.3
python-bidi==0.4.2
python-dateutil==2.8.2
python-docx==0.8.11
python-dotenv==1.0.0
python-json-logger==2.0.7
python-magic==0.4.27
python-pptx==0.6.21
python3-openid==3.2.0
pytz==2023.3.post1
pywin32==306
pywinpty==2.0.10
PyYAML==6.0
pyzmq==25.0.2
qrcode==7.4.2
qtconsole==5.4.2
QtPy==2.3.1
rcssmin==1.1.1
regex==2023.5.5
reportlab==4.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986==1.5.0
rfc3986-validator==0.1.1
rich==13.0.1
rjsmin==1.2.1
s3transfer==0.6.1
screen==1.0.1
seaborn==0.12.2
semantic-version==2.8.5
Send2Trash==1.8.2
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4.1
South==1.0.2
SQLAlchemy==2.0.18
sqlparse==0.4.4
stack-data==0.6.2
starlette==0.26.1
streamlit==1.22.0
streamlit-chat==0.0.2.2
svglib==1.5.1
sympy==1.12
tailwind==3.1.5b0
tenacity==8.2.2
termcolor==1.1.0
terminado==0.17.1
tiktoken==0.4.0
TimeConvert==3.0.13
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.8
toolz==0.12.0
tornado==6.3.1
tqdm==4.65.0
traitlets==5.9.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.0.1
unicodecsv==0.14.1
unstructured==0.6.6
uri-template==1.2.0
uritools==4.0.2
urllib3==1.26.18
uvicorn==0.22.0
validators==0.20.0
virtualenv==20.25.0
virtualenvwrapper-win==1.2.7
watchdog==3.0.0
watchfiles==0.19.0
wcwidth==0.1.9
weasyprint==60.1
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
websockets==11.0.3
Werkzeug==2.3.6
whitenoise==6.6.0
widgetsnbextension==4.0.7
wrapt==1.15.0
xhtml2pdf==0.2.15
XlsxWriter==3.1.0
xlwt==1.3.0
yarl==1.9.2
zipp==3.15.0
zopfli==0.2.3
zstandard==0.21.0
